New feature to provide private messages between members Fixes #84

+++ chat/views/views_public_rooms.py
@@ -1,24 +1,26 @@
 import logging
-from urllib.parse import unquote, urlencode
 from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import ValidationError
+from django.db.models import Count, OuterRef, Subquery
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
-from django.contrib.auth.decorators import login_required
-from django.core.exceptions import ValidationError
-from django.contrib import messages
-from django.db.models import Count, OuterRef, Subquery
+from django.utils.translation import gettext as _
 
+from members.models import Member
+from ..models import ChatMessage, ChatRoom
 from cousinsmatter.utils import Paginator, is_ajax
 from cm_main import followers
-from members.models import Member
-from .models import ChatMessage, ChatRoom
+
+from urllib.parse import unquote, urlencode
 
 logger = logging.getLogger(__name__)
 
 
 @login_required
-def chat(request, page_num=1):
+def chat_rooms(request, page_num=1):
   # Subquery to get the author of the first related ChatMessage instance of a room
   first_msg_auth_subquery = ChatMessage.objects.filter(
     room=OuterRef('pk')
@@ -29,7 +31,7 @@ def chat(request, page_num=1):
   ).values('member_id')
 
   # Annotate room instances with the first message and the number of messages in the room
-  chat_rooms = ChatRoom.objects.all().annotate(
+  chat_rooms = ChatRoom.objects.public().annotate(
     num_messages=Count("chatmessage"),
     first_message_author=Subquery(first_msg_auth_subquery)
     ).order_by('date_added')
@@ -57,7 +59,11 @@ def chat(request, page_num=1):
 def new_room(request):
   room_name = unquote(request.GET['name'])
   try:
-    room, created = ChatRoom.objects.get_or_create(name=room_name)
+    from django.utils.text import slugify
+    slug = slugify(room_name)
+    room, created = ChatRoom.objects.get_or_create(slug=slug)
+    if not room.is_public():
+      raise ValidationError(_("A private room with almost the same name already exists: %s") % room.name)
     room_url = reverse('chat:room', args=[room.slug])
     # if room was created, check user followers
     if created:
@@ -74,7 +80,7 @@ def new_room(request):
           pass
         case _:
           messages.error(request, f'{error[0]}: {" ".join(error[1])}')
-    return redirect(reverse('chat:chat'))
+    return redirect(reverse('chat:chat_rooms'))
 
 
 @login_required
@@ -103,6 +109,8 @@ def toggle_follow(request, room_slug):
 def edit_room(request, room_slug):
   if is_ajax(request):
     room = get_object_or_404(ChatRoom, slug=room_slug)
+    if request.user != room.owner():
+      raise ValidationError(_("Only the owner of a room can edit it"))
     # print(request.POST)
     if 'room-name' not in request.POST:
       raise ValidationError("No room name provided")
@@ -116,23 +124,9 @@ def edit_room(request, room_slug):
 @login_required
 def delete_room(request, room_slug):
   room = get_object_or_404(ChatRoom, slug=room_slug)
+  if request.user != room.owner():
+    messages.error(request, _("Only the owner of a room can delete it"))
+    return redirect(reverse('chat:chat_rooms'))
   room.delete()
   messages.success(request, f"Chat room {room.name} deleted")
-  return redirect("chat:chat")
-
-
-@login_required
-def create_test_rooms(request, num_rooms):
-  for i in range(num_rooms):
-    ChatRoom(name=f"a chat room #{i}").save()
-  return redirect("chat:chat")
-
-
-@login_required
-def create_test_messages(request, num_messages):
-  room = ChatRoom.objects.create(name="A chat room for testing a lot of messages")
-  connected_member = Member.objects.get(id=request.user.id)
-  for i in range(num_messages):
-    ChatMessage(room=room, member=connected_member,
-                content=f"a chat message #{i}").save()
-  return redirect("chat:room", args=[room.slug])
+  return redirect('chat:chat_rooms')
